requirements.txt
ultralytics
opencv-python
pytesseract
mss
numpy
streamlit
python-dotenv
onepassword-sdk
Pillow